Added division tests and ensured they failed. Also rewrote the division declatation to be float instead of double to work with tests

+++ calculator.h
@@ -3,5 +3,5 @@
 int add(int a, int b);
 int subtract(int a, int b);
 int multiply(int a, int b);
-double divide(int a, int b);
+float divide(int a, int b);
 #endif
